# src/backend/internal/platform/postgres/dbgen/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.31.1

package dbgen

import (
	"time"

	"github.com/jackc/pgx/v5/pgtype"
)

type AnalysisMeterResult struct {
	AnalysisRunID     int64
	MeterID           string
	ComputedStatus    string
	EvaluatedReadings int32
	FlaggedReadings   int32
	FindingsCount     int32
}

type AnalysisRun struct {
	ID                  int64
	Status              string
	Stage               string
	ProgressPercent     int16
	EngineVersion       string
	Configuration       []byte
	MetersCount         *int32
	ReadingsCount       *int32
	EventsCount         *int32
	FindingsCount       *int32
	HighPriorityCount   *int32
	AggregateConfidence *float64
	ErrorCode           *string
	ErrorMessage        *string
	CreatedAt           time.Time
	StartedAt           *time.Time
	CompletedAt         *time.Time
}

type Anomaly struct {
	ID                      int64
	AnalysisRunID           int64
	MeterID                 string
	Priority                int32
	Type                    string
	Severity                string
	Confidence              float64
	Rule                    string
	RecommendedAction       string
	Reason                  string
	StartedAt               time.Time
	LastObservedAt          time.Time
	DurationSeconds         int64
	ConsumptionDeviationPct float64
	Status                  string
	Evidence                []byte
	CreatedAt               time.Time
}

type Event struct {
	ID             int64
	MeterID        string
	EventTimestamp time.Time
	EventType      string
	Description    string
}

type Meter struct {
	MeterID string
}

type Reading struct {
	MeterID          string
	ReadingTimestamp time.Time
	ConsumptionKwh   pgtype.Numeric
	VoltageV         pgtype.Numeric
	CurrentA         pgtype.Numeric
	PowerFactor      pgtype.Numeric
	SourceStatus     string
}